producer/pyproject.toml
[project]
name = "producer"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "confluent-kafka",
    "python-dotenv"
]
